+++ src/fractal_client/auth/__init__.py
@@ -1,0 +1,1 @@
+from ._client import AuthClient  # noqa: F401
