+++ src/fractal_client/cmd/_aux_task_caching.py
@@ -4,7 +4,7 @@ from typing import Any
 
 import packaging.version
 
-from fractal_client.auth.authclient import AuthClient
+from fractal_client.auth import AuthClient
 from fractal_client.config import settings
 from fractal_client.response import check_response
