+++ src/fractal_client/cmd/_profile.py
@@ -1,7 +1,7 @@
 import json
 from pathlib import Path
 
-from fractal_client.auth.authclient import AuthClient
+from fractal_client.auth import AuthClient
 from fractal_client.interface import Interface
 from fractal_client.response import check_response
