+++ src/fractal_client/cmd/_project.py
@@ -1,4 +1,4 @@
-from fractal_client.auth.authclient import AuthClient
+from fractal_client.auth import AuthClient
 from fractal_client.interface import Interface
 from fractal_client.response import check_response
