+++ src/fractal_client/cmd/_task_collection.py
@@ -4,7 +4,7 @@ import sys
 from pathlib import Path
 from typing import Literal
 
-from fractal_client.auth.authclient import AuthClient
+from fractal_client.auth import AuthClient
 from fractal_client.interface import Interface
 from fractal_client.response import check_response
